requirements.txt
sentence-transformers==2.7.0
faiss-cpu==1.8.0
groq==0.11.0
httpx==0.27.2
flask==3.0.3
flask-cors==4.0.1
numpy==1.26.4